RAiD-270 - Re-organised project structure and updated components & containers mapping
