+++ raid-agency-app/src/components/breadcrumbs-bar/index.tsx
@@ -1,2 +1,2 @@
 export { BreadcrumbsBar } from "./BreadcrumbsBar";
-export type { Breadcrumb } from "./types";
+export type { Breadcrumb } from "../../generated/raid/Breadcrumb";
