+++ raid-agency-app/src/pages/raid-display/components/RaidDisplayMenu.tsx
@@ -1,5 +1,5 @@
-import { InviteButton } from "@/components/invite/InviteButton";
-import { InviteDialog } from "@/components/invite/InviteDialog";
+import { InviteButton } from "@/containers/invite/InviteButton";
+import { InviteDialog } from "@/containers/invite/InviteDialog";
 import {
   Edit as EditIcon,
   KeyboardArrowUp as KeyboardArrowUpIcon,
